Add bootstrap theme and rewrite models-views
Bootstrap theme was added to project. There are still many things to make smooth
but for now the main topic will be deployment and adding images, thus themes were
prepared to handle thumbnails.
In  paralell, tests were rewritten at large. That enforced changes in models and views.
Majority of the logic was moved to models, and tests should cover them thoroughly
(apart from authorization part). Tests for views are still to be finished.